Change CONFIG_UPLOAD_MESSAGE to avoid confusion
Changed the CONFIG_UPLOAD_MESSAGE to make it clear that, to upload a previous map, you have to upload a "single" JSON that contains "both" the old data and all config. 

Trying to upload separate JSONs of previously used config and previously used data, which seems like a possible way forward with the old message, in fact doesn't work and results in a map that tries to load forever. 

This small change makes it clear that users should look for the option that exports both the config and data together in a single JSON. This is a radio button on the bottom of the export window that changes from the default state of separate exports to an 'everything' export.