Decrease filter step size for small domains
Otherwise the detected filter step size may be larger than the domain,
which results in not being able to use the filter. This is applicable
when working with small percentages (<0.1%).

While changing this I also noticed this will treat negative numbers as
very small; I changed `getNumericStepSize` to compute the step size
based on absolute value.

Signed-off-by: Isaac Brodsky <isaac@isaacbrodsky.com>