[fix] correct provider downloadMap type
This fixes the return type of the `downloadMap()` method in `Provider`.

It was `Promise<{ map: SavedMap; format: string }>`, but `SavedMap` itself has a `map` property. When an implementation of `downloadMap()` returns an object like `{ map: { map: { ... } }, format: ... }`, loading maps works, but the "Save"/"Save As" functionality breaks. I haven't figured out why, but I tested manually and confirmed that flattening the nested `map` property fixes the re-saving, without breaking the loading.

Signed-off-by: Eric Peterson <git@ericp.co>